chore: misc PR review updates
