+++ packages/slice-machine/src/components/Card/Card.module.css
@@ -209,8 +209,7 @@
   display: flex;
   align-items: center;
   justify-content: center;
-  width: 32px;
-  height: 32px;
+  padding: 8px;
   border: 1px solid var(--grey6);
   border-radius: 6px;
 }
